+++ thoth-api/migrations/v0.13.4/down.sql
@@ -1,0 +1,3 @@
+DROP TABLE contact_history;
+DROP TABLE contact;
+DROP TYPE contact_type;